vintage.py has been moved one directory up to address the comment in #121
