+++ src/example.py
@@ -46,11 +46,11 @@ from csv2df.specification import Definition, Specification, SPEC
 from csv2df.reader import Reader, open_csv
 from csv2df.parser import extract_tables
 from csv2df.emitter import Emitter
-from csv2df.vintage import Vintage
+from vintage import Vintage
 from csv2df.validator import Validator
 
 
-# copied from csv2df.runner.py
+# copied from runner.py
 #HS: csv2df.runner.py changed to csv2df.vintage.py
 
 def get_dataframes(csvfile, spec):
